task/BS-02-32: Add sorted feature specs
