+++ spec/features/books/books_list_spec.rb
@@ -1,5 +1,5 @@
 RSpec.describe 'BooksList', type: :feature do
-  let(:catalog_page) { Pages::Catalog.new }
+  let(:catalog_page) { Pages::Books::Catalog.new }
 
   before do
     create_list(:book, 2)
