+++ spec/rails_helper.rb
@@ -5,8 +5,8 @@ abort('The Rails environment is running in production mode!') if Rails.env.produ
 require 'rspec/rails'
 
 Dir[Rails.root.join('spec/support/config/*.rb')].each { |file| require file }
-Dir[Rails.root.join('spec/support/sections/*.rb')].each { |file| require file }
-Dir[Rails.root.join('spec/support/pages/*.rb')].each { |file| require file }
+Dir[Rails.root.join('spec/support/sections/**/*.rb')].each { |file| require file }
+Dir[Rails.root.join('spec/support/pages/**/*.rb')].each { |file| require file }
 
 begin
   ActiveRecord::Migration.maintain_test_schema!
