+++ spec/support/sections/books/books_list.rb
@@ -1,0 +1,7 @@
+module Sections
+  module Books
+    class BooksList < SitePrism::Section
+      elements :books, '.book'
+    end
+  end
+end
