+++ spec/support/sections/books/sorting.rb
@@ -1,0 +1,8 @@
+module Sections
+  module Books
+    class Sorting < SitePrism::Section
+      elements :categories_links, "ul.list-inline li:nth-last-child(-n+#{BookDecorator::NEWEST_BOOKS_COUNT})"
+      elements :sorting_links, 'div.dropdown > ul.dropdown-menu li'
+    end
+  end
+end
