--- spec/support/sections/books_list.rb
@@ -1,5 +1,0 @@
-module Sections
-  class BooksList < SitePrism::Section
-    elements :books, '.book'
-  end
-end
