--- spec/support/sections/slider.rb
@@ -1,5 +1,0 @@
-module Sections
-  class Slider < SitePrism::Section
-    elements :buy_now_buttons, 'div.general-align>button.btn-primary'
-  end
-end
